refactor: user-info module
