+++ src/storyblok-auth-api/index.ts
@@ -1,7 +1,4 @@
-export * from './user-info/UserInfo/user-info'
 export * from './auth-handler'
-export { User } from './user-info/User/user'
-export { Space } from './user-info/Space/space'
-export { Role } from './user-info/Role/role'
-export { AuthHandlerParams } from './AuthHandlerParams'
-export { refreshToken } from './refreshToken'
+export * from './user-info'
+export * from './AuthHandlerParams'
+export * from './refreshToken'
