+++ src/storyblok-auth-api/user-info/Role/index.ts
@@ -1,0 +1,3 @@
+export * from './Role'
+export * from './isRole'
+export * from './isRoles'
