+++ src/storyblok-auth-api/user-info/Role/isRole.ts
@@ -1,5 +1,5 @@
 import { hasKey } from '../../../utils'
-import { Role } from './role'
+import { Role } from './Role'
 
 export const isRole = (obj: unknown): obj is Role =>
   hasKey(obj, 'name') && typeof obj.name === 'string'
