+++ src/storyblok-auth-api/user-info/Space/index.ts
@@ -1,0 +1,2 @@
+export * from './Space'
+export * from './isSpace'
