+++ src/storyblok-auth-api/user-info/User/index.ts
@@ -1,0 +1,2 @@
+export * from './User'
+export * from './isUser'
