+++ src/storyblok-auth-api/user-info/User/isUser.ts
@@ -1,5 +1,5 @@
 import { hasKey } from '../../../utils'
-import { User } from './user'
+import { User } from './User'
 
 export const isUser = (obj: unknown): obj is User =>
   hasKey(obj, 'id') &&
