+++ src/storyblok-auth-api/user-info/UserInfo/UserInfo.ts
@@ -1,9 +1,9 @@
 /**
  * The data that is returned from https://app.storyblok.com/oauth/user_info
  */
-import { User } from '../User/user'
-import { Space } from '../Space/space'
-import { Role } from '../Role/role'
+import { User } from '../User'
+import { Space } from '../Space'
+import { Role } from '../Role'
 
 export type UserInfo = {
   user: User
