+++ src/storyblok-auth-api/user-info/UserInfo/index.ts
@@ -1,0 +1,2 @@
+export * from './UserInfo'
+export * from './isUserInfo'
