+++ src/storyblok-auth-api/user-info/UserInfo/isUserInfo.test.ts
@@ -1,7 +1,7 @@
 import { isUserInfo } from './isUserInfo'
-import { User } from '../User/user'
-import { Space } from '../Space/space'
-import { Role } from '../Role/role'
+import { User } from '../User'
+import { Space } from '../Space'
+import { Role } from '../Role'
 
 const user: User = {
   id: 123,
